MSVC++ YACoin additions to 40 source files
These additions do not modify the original sources for the production of
YACoind and YACoin-qt.

A new file in the doc directory, called Video scripts and notes.txt
details DLL building and MSVC++ configuring needed for a correct build.
I'm working on videos to detail this process for those that may need
more explanation.