show library versions, moving wallets, live splash screen
All of these 044 changes can be see in this video:
http://youtu.be/C8GvHpjbAnM
which was done to the 043 code. In addition:
In YACoin.conf (or on the command line) for Windows (#ifdef WIN32)
versions of YACoind (MS), YACoind (gcc) & YACoin-qt (gcc) there is a new
walletpath argument.

walletpath can take a full or partial path (or just a filename) to an
existing 'wallet.dat' file or a non-existing name to create a new wallet
with that new name.  If it's a partial path or just a name, the code
'falls back' to using datadir to compose the full path, which is what
the normal code does anyway.

The 'wallet.dat' name can be any legal filename!

The kinds of things one can do with this capability are: Run the wallet
off of a USB memory drive and never have it on the hard drive of the
computer!

Change wallets by editing a line in YACoin.conf, without losing the
other wallet, which is the same as running the same blockchain with
different wallets.

Perhaps an odd use of running the same wallet on different blockchains?

All one needs to know is how the BerkeleyDB code is written in
Bitcoin/YACoin, etc.

In YACoin, the wallet code expects a blkindex.dat and (possibly) a
database/log.0000000nnn file in the directory with the wallet file,
'wallet.dat' or whatever you want to rename it!

The blockchain, I think, is in the datadir directory, wherever that is
on your computer, and it expects to find a blk0001.dat and perhaps a
peers.dat and the ability to have a .lock file.

There is some connection, I believe, but again there is no
documentation, on the relationship between the 'datadir'/blk001.dat file
and the blkindex.dat and (possibly) a database/log.0000000nnn files in
the walletpath which now can be a separate directory!

Ron