Built in explorer added to YACoin-Qt!
Added a "mode-less" block chain explorer to YACoin-Qt.  Added some
"missing" close buttons to a few dialogs.  Pretty-ed up the About
dialog,  Made the "selected" text visible!  Blocked the ability to play
around in Qt before the splash screen ends, since it can cause Qt to
crash, in Windows at least!  Centered the text in the splash screen.
One icon and two source files and three Qt form .ui files added..  Not
tested on Mac or linux, but easy enough, see bitcoingui.cpp

This is a first pass as it were to show how easy it is to add this
functionality.  Modify it to show hat you want to see.  Note the
tooltips, shortcuts and other subtle "touches" made to make the Qt
"experience" more SAA CUA compatible.