Changes to stabilize the Qt version and fix random crashes
Some of the things fixed in this set of changes:

Errors in the structure of the wallet.h code that caused random
exceptions have been fixed.

Errors in the getblock series of RPC commands that prevented external
mining have been fixed.

The Qt build has been improved for various OSs, so that the Qt version
now runs.

Things that have been discovered and need to be fixed, and will be soon
are:

The HTTPS price APIs need to be done.  This means that now 'getyacprice'
no longer
works since all web HTTP APIs are now only HTTPS.

Also the RPC calls built into the QT version Debug window -> Console
need to be fixed with respect to
the RPC architecture around try - catch.  What this means is that errors
in any command can cause an
exception that may or may not be properly handled by Qt.  If not
handled, then it aborts.

The daemon has no problem.

So the Qt version works, but the RPC console is 'risky'.

Fixes to the code for starting from the genesis block on the main net
(main.cpp) were by Artur.
The -rescan bypass speed up fixes were by Matjaz
The improper 'gethashespersec' and 'getmininginfo' results fixes were by
Ben.
The wallet.h irregular crash in getcredit fixes were by Matjaz.

Complete build time speedup of 25% + or - was done by virtue of doubly
guarded
include files.  See
https://gcc.gnu.org/onlinedocs/gcc-4.5.0/cpp/Once_002dOnly-Headers.html
where it states that gcc >= 4.5.x open but doesn't read past the
#ifndef.  Note that it
still has to open the file. I see improvements in gcc, less so in MSVC++
There is 'controversy here, see
https://google.github.io/styleguide/cppguide.html
where I like lots of things there, that are not done in any coin
sources!
and

https://stackoverflow.com/questions/2233401/are-redundant-include-guards-necessary
and
http://wiki.c2.com/?RedundantIncludeGuards
I will complete the guards on all the .h files that include other .h
files, since this is a large project.
It is mostly done.  I think that the .cpp files would benefit from this
too, so I did it on a few.

The real killer for YAC & all the other coins is the nested recursive
inclusions done by boost
with its .hpp files.  Now that would really speed things up if they were
all fixed!!

I 'lossened' the unneccessary strict casing of the RPC commands. RPC
arguments are still case sensitive.
Adjusted display constants at startup for a 'smoother' display.

the constant STRICT was colliding only in Windows, so changed to STRICT_

Ron