Port default reverse (#84)
Fix a change that has been done for development previously. The default port would return first the configured port and then the default port 7688. This creates a problem when trying to connect to other nodes, since if no port is specifically specified, the default Yacoin port should be used to establish a connection.